แก้ mat_type

+++ application/modules/product/views/product.php
@@ -16,23 +16,26 @@
     <div class="row">
         <form method="get">
             <div class="form-group">
-            <?php
-            foreach ($product_list as $key => $value){
-            ?>
-                <div class="col-md-3">
-                    <div class="thumbnail img-check">
-                        <center><a href= "<?php echo site_url ('product/product_edit/'.$value['mat_id']);?>"><label for="cb1"><img src="<?php echo base_url('images_compare/'.$value['mat_pic']);?>" alt="light" width="120" height="120"></label></a></center>
-                        <div class="caption">
-                            <center><p><?php echo $value['mat_name'];?> : <?php echo $value['mat_quantity'];?> <?php echo $value['mat_unit'];?> <?php echo $value['mat_cost'];?> บาท </p></center>
-                            <div class="col-sm-9">
+                <?php
+                foreach ($product_list as $key => $value){
+                    if ($value['mat_type']==1){
+                        ?>
+
+                        <div class="col-md-3">
+                            <div class="thumbnail img-check">
+                                <center><a href= "<?php echo site_url ('product/product_edit/'.$value['mat_id']);?>"><label for="cb1"><img src="<?php echo base_url('images_compare/'.$value['mat_pic']);?>" alt="light" width="120" height="120"></label></a></center>
+                                <div class="caption">
+                                    <center><p><?php echo $value['mat_name'];?> : <?php echo $value['mat_quantity'];?> <?php echo $value['mat_unit'];?> <?php echo $value['mat_cost'];?> บาท </p></center>
+                                    <div class="col-sm-9">
+                                    </div>
+                                    <a href="<?php echo site_url('product/delete_product/'.$value['mat_id']);?>" type="delete" class="btn btn-danger">ลบ</a>
                                 </div>
-                                <a href="<?php echo site_url('product/delete_product/'.$value['mat_id']);?>" type="delete" class="btn btn-danger">ลบ</a>
+                            </div>
                         </div>
-                    </div>
-                </div>
-            <?php
-            }
-            ?>
+                        <?php
+                    }
+                }
+                ?>
             </div>
         </form>
     </div>
